Last try with my implementation of kmeans.
The issue is: some clusters disappear (as if a _cluster outside_ a set of points is closer to the set than one point _in_ the set)
Other issue is that I didn't do it in flink, just scala, so it does not distribute. :(

Well, let's look for a implementation that works, like MOAs.