[WIP] Minimize classifier->detector talk...
classifier does not send non-intresting identified results, stores the ids and sends on next communication
Detector stores examples that were skipped (because interlacing of classifier threads)
and receives the non-intresting ids, looks for them in the skipped, then in the fligth controller buffer, moving from FC to Skipped if not in non-intresting.
That is a lot of work.